# packages/contracts/src/openapi.yaml
openapi: 3.1.0
info:
  title: Xenitex Appliance API
  version: 1.0.0
  description: >
    v1 of the on-premise appliance API. This document is the single source of
    truth (P1-01) — the TypeScript client and types in this package are
    generated from it, never hand-written (P1-02).

    Conventions (see docs/adr/0006-api-conventions.md, accepted at GATE 1):
    opaque cursor pagination on every collection (`?cursor=&limit=`, response
    `{ items, nextCursor }`); RFC 9457 Problem Details on every non-2xx
    response; `Idempotency-Key` required on every creation POST;
    `ETag`/`If-Match` required on every mutation of a mutable resource; URI
    path versioning (`/v1/...`); long-running operations return `202` with a
    `Location` header pointing at a pollable resource, never a blocking
    request.

    JSON casing: camelCase, matching packages/domain. Any field whose value
    originates from a scanned target (SEC-17/AIP-02 — banners, headers,
    certificate subjects, hostnames, identity-key values, OS/service/product
    inference) carries an `Untrusted` suffix in its property name. This is
    the same greppable-naming convention as the `_untrusted` column suffix in
    docs/adr/0007-schema-conventions.md, extended to the API boundary so a
    frontend consumer can never mistake such a value for first-party data —
    it must be escaped on output, never rendered as HTML, never interpreted.
servers:
  - url: /v1
...
security:
  - sessionCookie: []
paths:
...
  /saved-views:
    get:
      tags: [SavedViews]
      operationId: listSavedViews
      summary: List saved views
      responses:
        '200':
          description: 'OK.'
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/SavedView' } }
    post:
      tags: [SavedViews]
      operationId: createSavedView
      summary: Create a saved view
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/SavedViewCreate' } } }
      responses:
        '201':
          {
            description: 'Created.',
            content: { application/json: { schema: { $ref: '#/components/schemas/SavedView' } } },
          }
        '400': { $ref: '#/components/responses/ValidationError' }
  /saved-views/{viewId}:
    parameters:
      - name: viewId
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [SavedViews]
      operationId: getSavedView
      summary: Get a saved view
      responses:
        '200':
          {
            description: 'OK.',
            content: { application/json: { schema: { $ref: '#/components/schemas/SavedView' } } },
          }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [SavedViews]
      operationId: updateSavedView
      summary: Update a saved view
      parameters: [{ $ref: '#/components/parameters/IfMatch' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/SavedViewCreate' } } }
      responses:
        '200':
          {
            description: 'OK.',
            content: { application/json: { schema: { $ref: '#/components/schemas/SavedView' } } },
          }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/StaleResource' }
    delete:
      tags: [SavedViews]
      operationId: deleteSavedView
      summary: Delete a saved view
      responses:
        '204': { description: Deleted. }
        '404': { $ref: '#/components/responses/NotFound' }

  # ------------------------------------------------------------- Exceptions
...
  /reports:
    get:
      tags: [Reports]
      operationId: listReports
      summary: List reports
      parameters:
        [{ $ref: '#/components/parameters/Cursor' }, { $ref: '#/components/parameters/Limit' }]
      responses:
        '200':
          description: 'OK.'
          content:
            application/json:
              schema:
                type: object
                required: [items, nextCursor]
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/Report' } }
                  nextCursor: { oneOf: [{ type: string }, { type: 'null' }] }
    post:
      tags: [Reports]
      operationId: createReport
      summary: Generate a report (4.8) — executive summary, technical detail, or delta
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReportCreate' }
      responses:
        '202':
          description: 'Accepted.'
          headers: { Location: { schema: { type: string } } }
          content: { application/json: { schema: { $ref: '#/components/schemas/Report' } } }
        '400': { $ref: '#/components/responses/ValidationError' }
  /reports/{reportId}:
    parameters:
      - name: reportId
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [Reports]
      operationId: getReport
      summary: Get a report
      responses:
        '200':
          {
            description: 'OK.',
            content: { application/json: { schema: { $ref: '#/components/schemas/Report' } } },
          }
        '404': { $ref: '#/components/responses/NotFound' }
  /reports/{reportId}/download:
    parameters:
      - name: reportId
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [Reports]
      operationId: downloadReport
      summary: Download a completed report
      parameters:
        - name: format
          in: query
          required: true
          schema: { type: string, enum: [html, csv, json] }
      responses:
        '200':
          description: Report artifact.
          content:
            text/html: { schema: { type: string } }
            text/csv: { schema: { type: string } }
            application/json: { schema: { type: object } }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Report generation not yet complete.
          content:
            {
              application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } },
            }

  # -------------------------------------------------------------- Notifications
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: xenitex_session
      description: >
        Opaque server-side session (SEC-08). State-changing requests also
        require the double-submit CSRF token in the `X-CSRF-Token` header,
        matched against the session's csrfTokenHash server-side.
  parameters:
    Cursor:
      name: cursor
      in: query
      description: Opaque cursor from a previous page's `nextCursor`. Never a raw offset (ADR 0006).
      schema: { type: string }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 500, default: 50 }
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: Current ETag of the resource. A stale value returns 409 (ADR 0006).
      schema: { type: string }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Client-generated key. A retried request with the same key returns the original response rather than creating a second resource (ADR 0006).
      schema: { type: string }
...
  headers:
...
    CorrelationId:
      description: >-
        OPS-01 — the correlation identifier for this request, propagated from
        the HTTP request through any queue job and scanner invocation it
        triggers. Echoes an inbound `X-Correlation-Id` if the client sent
        one, otherwise server-generated. Surfaced on error responses so a UI
        can quote it back in a support conversation (see ErrorState).
      schema: { type: string }
  responses:
...
    ValidationError:
      description: Schema validation failed at the API boundary (SEC-12) — unknown fields are rejected, not ignored.
      headers: { X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' } }
      content:
        { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }
...
    NotFound:
      description: No such resource.
      headers: { X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' } }
      content:
        { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }
...
    StaleResource:
      description: 'code: concurrency.stale_resource — the supplied If-Match no longer matches (ADR 0006).'
      headers: { X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' } }
      content:
        { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }
...
  schemas:
    ProblemDetails:
      type: object
      required: [type, title, status, code]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code:
          type: string
          description: Stable, machine-readable, namespaced (e.g. `scope.exclusion_violation`). Clients branch on this, never on title/detail (ADR 0006).
...
    ReportTemplate: { type: string, enum: [executive_summary, technical_detail, delta] }
    ReportStatus: { type: string, enum: [pending, completed, failed] }
...
    SavedView:
      type: object
      required: [id, name, query, isShared]
      properties:
        id: { type: string }
        name: { type: string }
        query:
          type: string
          description: The encoded filter/sort/pagination query string (P1-25) — a saved view is a named, storable version of a shareable URL.
        isShared: { type: boolean }
        createdBy: { type: string }
    SavedViewCreate:
      type: object
      required: [name, query, isShared]
      properties:
        name: { type: string }
        query: { type: string }
        isShared: { type: boolean }
...
    ReportCreate:
      type: object
      required: [template, scopeFilter, formats]
      properties:
        template: { $ref: '#/components/schemas/ReportTemplate' }
        scopeFilter: { type: object, additionalProperties: true }
        dateRangeStart: { type: string, format: date, description: delta template only. }
        dateRangeEnd: { type: string, format: date }
        formats: { type: array, items: { type: string, enum: [html, csv, json] }, minItems: 1 }
    Report:
      type: object
      required: [id, template, status, formats, generatedBy, generatedAt]
      properties:
        id: { type: string }
        template: { $ref: '#/components/schemas/ReportTemplate' }
        scopeFilter: { type: object, additionalProperties: true }
        dateRangeStart: { oneOf: [{ type: string, format: date }, { type: 'null' }] }
        dateRangeEnd: { oneOf: [{ type: string, format: date }, { type: 'null' }] }
        status: { $ref: '#/components/schemas/ReportStatus' }
        dataVersions:
          oneOf:
            - type: object
              properties:
                vulnerabilityDataImportId: { type: string }
                riskScoringPolicyVersion: { type: integer }
            - type: 'null'
        formats: { type: array, items: { type: string, enum: [html, csv, json] } }
        blobStoreKey: { oneOf: [{ type: string }, { type: 'null' }] }
        generatedBy: { type: string }
        generatedAt: { type: string, format: date-time }
        completedAt: { oneOf: [{ type: string, format: date-time }, { type: 'null' }] }
...